Add HELIO mode.
We now distinguish between two modes of operation for NAPLES: OTM
(propagation with switch of primary bodies) and HELIO (heliocentric
propagation.
If n_rs = 0 NAPLES works in HELIO mode, in which no switch is
performed to the geocentric reference system. In this case, no output
is computed at the switch points, but this could be implemented in the
future since those points are still present in the reference
trajectory and in the times vector.
The HELIO mode has been validated with the CEsim-h code.

Correction in the input.txt title.

.gitignore has been temporarily modified as to delete unnecessary
files from the remote.