Add initial step size computation.
The initial step size is not constant anymore, and it depends on
the particular formulation (i.e. type of independent variable) used
and on the orbit characteristics.
For elliptic orbits, the initial step size is computed as
ds_i = 0.01*PER
where PER is the interval in the independent variable corresponding
to one orbital period in physical time in the unperturbed case.
For instance, in EDromo the independent variable reduces to the
eccentric anomaly in the unperturbed case, therefore PER = 2*pi.
For hyperbolic orbits, the initial step size is computed as
ds_i = (2*s+)/600
where s+ is the value of the independent variable corresponding to
a point on the unperturbed hyperbolic trajectory at a distance of
0.1 AU from the primary body, taken after the periapsis passage.
This criterion is conceptually equivalent to that routinely used
in asteroid impact monitoring: when a close encounter is detected
the integrator performs 600 steps in true anomaly in a hyperbolic
arc from r = -0.1 AU to r =  +0.1 AU, where r is the distance from
the primary body.