Fix the initial step size computation.
rlim = 0.1 au is too small for heliocentric orbits. If we have
hyperbolic heliocentric orbits, it is quite possible that the
perihelion will be r_peri > 0.1 au, so there is no value of true
anomaly which satisfies

cos(f) = (p/rlim - 1)/e

thus the computation of the initial step size in the DSTEPSIZE
function fails with a NaN.

We modified the initial time step computation to choose rlim = 5 au
when the orbit is heliocentric. Also, we added a rescaling of the
initial time step as a function of the integration tolerance as

ds_i = -ds_i/log10(tol).