Patch out-of-bounds bug if a desynchronization occurs.
In some very rare cases (~once each 8 million propagations), the
root-finding used in phase CE to synchronize the test and reference
trajectories at the same values of physical time fails for a reason
yet unknown.
This results in dp_CE not having the right length for storage in
dp_full, thus raising an out-of-bounds runtime error.
We introduced a sanity check which skips the current iteration
in case in which this condition is detected.