Add fix to the previous patch.
The patch in the previous commit did not allow any propagation to
take place when in HELIO mode (everything was discarded with
istate = -13 since len_H1 + len_CE + len_H2 - 2 /= lenref in HELIO).
Therefore the safety check is not executed when n_rs == 0.
N.B: we could have executed the safety check
len_H1 == lenref
in HELIO, but we chose not to. Therefore if the same problematic
behaviour arises in HELIO we will have more information to debug
it.