Several improvements and bugfixes.
LSODAR: Turned off the check on excessive accuracy for the LSODAR
solver. This allows specifying tolerances < 2.2e-16 in double
precision (may be useful for regularized formulations).

NAPLES.f90, integration.f90, makefile: Add the computation of the
average ratio between the number of function evaluations due to
root-finding and due to stepping for the xra15 integrator.

io.f90: Fixed a user interface bug.

xeverhart.f90: Fixed a bug in which for some specific cases Newton's
method for root-finding would not be initialized correctly.